Server with Transport Layer Security.
Change Connection::handleData to use `fread` once again so that, when it
is appropriate,  the data is filtered by an ssl/tls stream wrapper.

Add a third argument to Server::listen which should be an array which
can be passed directly to stream_context_create.  If the context has an
'ssl' key then it will stream_socket_enable_crypto for connected
sockets.