Improve TLS
Avoid needing to set the master server socket to block at the start of
new secure connections by moving the crypto enablement from Server into
a new SecureConnection. The implementation is based on [react-119] and
with feedback from @CriseX.

Revert changes to Connection that were made in the previous commit.

Improve the selection of a default set of SSL/TLS protocols to use.

[react-119]: /reactphp/react/pull/119