Do not expose $master property
Interfering with this (previously public) property could introduce some
very subtle bugs, so it's best to simply avoid exposing it.

Empirical evidence seems to suggest this isn't used much outside of this
package anyway.

Inside this package, we can safely replace all references to it with
references to child sockets created upon connection.