Avoid possibility of missing remote address when TLS handshake fails
This fixes a race condition where the remote address could appear to be
"emtpy" in the server exception message, that could occur when the
connection to the peer was already closed during the TLS handshake. This
is simply avoided by getting the remote address prior to starting the
TLS handshake, in which case the remote address is known to exist.