Work around failing test case detecting EOF on TLS 1.3 socket streams
This PR improves the test suite to avoid a possible race condition for
our TLS tests. It does not change anything about the actual behavior or
the expected output, but it helps making the expected output more
explicit and no longer subject to a possible race condition. This helps
avoiding possible false negatives if TLS 1.3 is supported and PHP
reports the EOF indicator before consuming all application data.

This builds on top of https://github.com/reactphp/socket/pull/185 and
https://github.com/reactphp/socket/pull/186