Avoid uneeded fragmented TLS work around for PHP 7.3.3+
The work around was introduced in order to bring TLS 1.3 support to all
supported versions, but at the same time it may cause very large data
chunks for high throughput scenarios. The underlying bug has been fixed
in PHP 7.3.3 and PHP 7.2.15, so we can avoid this now uneeded work
around on said version.