Improve error reporting when DNS lookup fails (happy eyeballs)
Also fixes a fatal error with legacy PHP 5 where it would fail with a
(catchable) fatal error when either DNS lookup fails. During test runs,
this would automatically be turned into an Exception and would
successfully reject the promise. Without an appropriate error handler, a
soft DNS error (such as when IPv6 is not available) would previously
terminate the program with a fatal error.