Change default socket backlog size to 511
The `backlog` parameter used to default to just 32 as of PHP 5.3.3 and
used to be hard coded to 4 in earlier PHP versions.

The new default of 511 was evaluated to be a reasonable value for
event-driven applications that can cope with many concurrent connections
on a variety of platforms. In particular, this solves common ECONNRESET
(connection reset by peer) errors when running benchmarks with higher
concurrency. For more specific requirements, the `backlog` parameter can
be given explicitly to override this default.

See https://github.com/php/php-src/commit/5b277c4da0d9fb8bfb59186b61a884aaafffb861